REFACTOR. Rename and reorganize the components 'CommandList' and 'CommandItem'
- Do not use the letter 's' for the plural, a letter can go missing, it is less confusing to use the word 'list':
Rename prop and data 'commands' to 'commandList'.
Rename data 'commandsFormatted' to 'commandListFormatted'.
Rename computed function 'getCommands' to 'getCommandList'.
- Unify the names with their components:
Rename the file and name component from 'ListCommands' to the 'CommandList'.
Rename class 'list-commands' to 'command-list'.
Rename class 'command' to 'command-item'.
Rename prop 'commandData' to 'commandItem'.
On the ‘v-for’ use the generic word ‘item'.
- Reorganize the components:
Move the components 'CommandList' and 'CommandItem' to the 'Command' folder.