REFACTOR. Rename and reorganize the components 'ReferenceList' and 'ReferenceItem'
- Do not use the letter 's' for the plural, a letter can go missing, it is less confusing to use the word 'list':
Rename prop 'references' to 'referenceList'.
- Unify the names with their components:
Rename the file and name component from 'ListReferences' to the 'ReferenceList'.
Rename the file and name component from 'Reference' to the 'ReferenceItem'.
Rename class 'list-references' to 'reference-list'.
Rename class 'reference' to 'reference-item'.
Rename prop 'dataReference' to 'referenceItem'.
On the ‘v-for’ use the generic word ‘item'.
- Reorganize the components:
Move the components 'ReferenceList' and 'ReferenceItem' to the 'Reference' folder.