BUGFIX. Disable some VSC settings
Reason:
- By default, the VSC configuration has `editor.wordWrap` and `editor.fontLigatures` enabled. But specifically for this project I want to disable them, so I don't have to modify the overall configuration.
These changes aim to improve code readability and maintain compatibility across different development contexts.

Steps:
- Disable line wrapping (`editor.wordWrap`):
Automatically splits lines of code or text to fit the width of the window or the maximum set.
This can be useful for reading or editing long code or text, as it is not necessary to scroll horizontally to see the entire contents.
Although it may split lines of code in unwanted places, which could affect the readability or visual organization of the code.
- Disable font ligatures (`editor.fontLigatures`):
A typographical feature that combines two or more adjacent characters (`->`, `!==`, `==`) into special symbols, which can be visually appealing.
This may not be supported in all development tools or programming environments, which may cause visual or rendering problems.
Also disabling this feature ensures that the developer will see the characters as they are, which may be preferable if they are more accustomed to the traditional visual representation of each character than to typographic ligatures.